hotfix for relative path in slides.
Redirect live and admin to live/ and admin/ which fixes their relative path.
This is not a final solution. A middleware will be use to redirect the links in the future (hopegully).
Also this does not fix the css for Margaritas work which seems to not fit with the templates.